Extract externalYamlEditor into modules/ (#1334 Step 9 direction)
The external YAML editor was added inline to 025-libraries.js in
#1659, adding ~985 lines to a file #1334 Step 9 already targets for
decomposition. This carves it out into
static/local-js/modules/externalYamlEditor.js as a
dependency-injected, idempotent init module -- the same pattern the
existing accordionHighlights/separatorPreview/validationPageBase
modules use.

Behavioural change: none. The block moved verbatim; only the
externalYamlEditorKinds table (which references per-page helpers like
setMetadataFileStatus and syncMetadataFilesEditor) and three other
cross-file dependencies (getActiveConfigName,
setLibrariesButtonPersistentBusy, prepareLibrariesModal) are now
injected via initExternalYamlEditorSubsystem({...}). The four
document.addEventListener calls are attached from inside the
(idempotent) init.

Net effect on 025-libraries.js: 9763 -> 8854 lines. New module: 972
lines. Added a vitest smoke test that pins the export contract.
Existing suites (1547 JS tests, tests/test_external_yaml_editor.py)
pass unchanged.
